Refactor: move random scheduler to its own module under scheduler
